+++ examples/react-example/CHANGELOG.md
@@ -1,5 +1,33 @@
 # react-example
 
+## 2.8.33
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/react@2.8.33
+
+## 2.8.32
+
+### Patch Changes
+
+- @stackframe/react@2.8.32
+
+## 2.8.31
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/react@2.8.31
+
+## 2.8.30
+
+### Patch Changes
+
+- @stackframe/react@2.8.30
+
 ## 2.8.29
 
 ### Patch Changes