+++ packages/template/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @stackframe/stack
 
+## 2.8.33
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.33
+  - @stackframe/stack-sc@2.8.33
+  - @stackframe/stack-ui@2.8.33
+
+## 2.8.32
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.32
+  - @stackframe/stack-ui@2.8.32
+  - @stackframe/stack-sc@2.8.32
+
+## 2.8.31
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-sc@2.8.31
+  - @stackframe/stack-shared@2.8.31
+  - @stackframe/stack-ui@2.8.31
+
+## 2.8.30
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.30
+  - @stackframe/stack-ui@2.8.30
+  - @stackframe/stack-sc@2.8.30
+
 ## 2.8.29
 
 ### Patch Changes